Ensure page is loaded on History.replaceState changes
Fixes an issue where router pages weren't being loaded when replaceState
was being used. Also simplifies the code a bit and ensures listeners are
removed when the router component is disconnected from the DOM.

Upgrades tests to not mock out pushState/replaceState which is
unnecessary.

Oh and upgrades sinon package which was pretty outdated.